Add item "null" into the ignore file for Git
The file null will be used as the error output for the Makefile. In the clean
command, we remove all the *.o and *.a file. But if there is non, there will
be an error information shown on the screen, which is quite troublesome. So we
redirect the error information to this file, which should not be tracked by
Git.