Found the source of the unstable phenomenon
1. The oscillations occur for the accuracy tests are caused by
   the bad linearization in the acoustic codes in the GRP solver.
2. There are still bugs lying in the newly inserted decomposition
   and recomposition codes. Now we use the old ones temporarily.
   Further carefull check is required.

We newly found that unstable phenomenon occurs in the accuracy test.
After careful check we can be sure that the source of the unstable
phenomenon lies in the acoustic part of code in the GRP solver. The
unproper linearization in the acoustic code cause the oscillations.

After finishing programming the ADER solver, we use the ADER codes to
replace the original acoustic codes from Prof. Li. But now it seems
that ADER codes would cause oscillations.

After use back the original acoustic code from Prof. Li, the numerical
results can be stable until m=320. But the oscillation occurs for the
case m=640. If we unconditionally use the exact Riemann solver every-
where, the numerical results can be stable until m=640.

Further, the newly inserted decomposition and recomposition codes still
have some bugs. Further carefull check is required.