Begin writing the dual solver for the RK4 scheme
   1. Begin writing the DUAL Riemann solver
   2. The oscillations in the accuracy tests have been
      eliminated

In Shu's SSP book, he told us that if the SSP 4th order RK scheme is
desired, negative flux coefficients b_ij would be inevitable. And for
the negative ones, DUAL approximated Riemann solver should be applied
(there is totally NO such a thing called DUAL EXACT RIEMANN SOLVER !).

So far, our RK scheme just uses normal Riemann solver everywhere, which
cannot enssure the SSP property for the scheme. Yet no unstable pheno-
menon has been obseved.

But for the consistency of the codes, we should add such a DUAL solver.

The second thing is that the oscillations in the accuracy test have been
eliminated by using normal time steps. Let me give you a explanation.

In order to debug, I compare the computation procedure with the one per-
formed by the MATLAB codes which are believed to have on bug. In the MAT-
LAB codes, I just set the time step by
                   tau = CFL*h
which is used in the computation for the scalar IVP
                 u_t + u_x = 0
		 u(x,0) = 1 + 0.2*sin(pi*x)
since the speed of the signal is just 1 in that case. In order to compare
the computation procedure, I also use the above time step, which will
cause no big trouble in the computation for the advection flow for the
Euler equations with the following initial condition
                 rho(x,0) = 1 + 0.2*sin(pi*x)
		 u(x,0)   = 1
		 p(x,0)   = 1
But from the previous commit we know that the solution is unstable for the
'm=640' case. Now I have correct the set for the time step by
          tau = CFL*h / sigma
where sigma computed by
                   sigma = max_j{|u_j+c_j|, |u_j-c_j|}
is the speed for the fastest characteristic signal. And
              +---------------------------------------+
	      |  THIS CURES THE UNSTABLE PHENOMENON!  |
              +---------------------------------------+
Now the accuracy test can be passed.