Finish three tasks
    1. Make the decomposition and recomposition in the reconstruction
       procedure a function.
    2. Finish check the accuracy test for the G4H5 scheme.
    3. Write the dual Roe solver for the SSP RK time discretization.

The first task has nothing special, just modify the function used in the
2-D case.

As for the second task, the unstable phenomenon observed in the accuracy
tests is caused by the unproper linearization in the GRP solver. It seems
that Prof. Li's original code is the best.
To find the reason still requires more works.

As for the third task, the tests for accuracy and the problems with dis-
continuities show no error.
After applying the dual flux in the RK scheme, the numerical errors for
the smooth problem decrease remarkably. But the resolution for the dis-
continuities also decreases remarkably. This may be cause by the increa-
sed dissipation.