Test the Z-type WENO and HWENO scheme
Now we have finish the Z-type RK-WENO and GRP-HWENO scheme. Both of them
have passed the accuracy tests. And computations upon problems with dis-
continuities has been done.

The Z scheme can remarkably lower the errors for the smooth problem by 1
magnitude for both the RK and the GRP scheme.

The Z scheme can improve the resolutions in the Shu-Osher problem for the
RK scheme. It can provide no help in the computation for the large pre-
ssure ratio problem as well as for the Woodward-Colella problem for the RK
scheme.

As for the GRP scheme, the improvement is also slight.