Set the thickness in the THINC reconstruction a parameter
We found that the primitive reconstruction should be used. And the fact
affect the stability is the thickness.

The larger it is, the thinner the reconstructed wave will be. But as
the wave get thinner, unstable phenomena would occur. 1.6 is recom-
manded in the original paper. But we found that a smaller value sh-
ould be applied to coopperate with the GRP fluxes.