Test some parameters of the entropy tube
The use of the reconstruction for conservative variables and the use
of the reconstruction for primitive variables show no difference.

And the use of the reconstruction for primitive variables gives a even
slightly better result of the numerical convergence orders.

Whether use the van Leer limiter in the computation shows no differen-
ce.

To bound the slope of the pressure (energy, resp.) to the slop of the
density is vary important in the computations.

The result is still not satisfactory. The L1 orders of [prim-
-nolimiter-bound] is '1.96, 2.28, 1.90, 1.17, 0.54, 0.13, -0.08'. And
the L-infinity orders are '1.42, 1.47, 1.53, 1.78, 1.37, 0.19, -0.36'.

The numerical orders for the moving boundary computations are even
worse.