The programm passes the accuracy tests
Including the advection of the desity and the entropy tube. We finally
find that the problem lies in the GRP solver.

The exact Riemann solution must be used in the acoustic cases. And all
the auxillary codes are not necessary anymore.

We do not need to bound the slope of the pressure to that of the density
anymore.